drf_proj/requirements.txt
asgiref==3.9.1
Django==5.2.5
django-jet-reboot==1.3.10
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-api-python-client==1.4.1
httplib2==0.31.0
oauth2client==4.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
pyparsing==3.2.4
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
uritemplate==4.2.0
